tui: setup main function
Now ready for tui development to begin! There are still several missing
features from the lib, but enough is there to start building a tui, and
having that built along side the lib will help me focus on important
things and make the lib as easy to use as possible.